AndroidIA: Maintain a refcount of open handles.
We can have multiple clients open the gralloc handle in same process.
We maintain a static pointer for Gralloc driver and we should
ensure that it is not closed till all handles are closed. Now,
we maintain a refcount of open handles and close driver only after
all handles get closed.

Jira: None.
Test: No crashes when secondary display is simulated or re-opening
      video applications.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>